requirements.txt
Flask==2.0.3
requests==2.28.1
psycopg2-binary==2.9.5
elasticsearch==8.10.1
Werkzeug==2.0.3
flask-apscheduler
